make ImportedInfections header-only
