+++ model/CMakeLists.txt
@@ -24,7 +24,6 @@ set (Model_CPP
   Host/Human.cpp
   Host/InfectionIncidenceModel.cpp
   Host/NeonatalMortality.cpp
-  Host/ImportedInfections.cpp
   
   Clinical/ClinicalModel.cpp
   Clinical/EventScheduler.cpp
